[FIX] 게시글 파일 업로드 불가 해결 #73 (#77)
* [feat] #73 게시글에 파일 추가, 수정 기능 추가

* [fix] #73 Test 진행 후, 의도와 다르게 동작하는 부분 수정

* [fix] 불필요한 import문 삭제

* [feat] #73 게시글에 파일 추가, 수정 기능추가

* [fix] 불필요한 import문 삭제

+++ src/main/java/com/abcdedu_backend/post/controller/CommentController.java
@@ -30,8 +30,8 @@ public class CommentController {
             @ApiResponse(responseCode = "404", description = "해당 사용자를 찾을 수 없습니다.", content = @Content)
     })
     @GetMapping("/members/{memberId}")
-    public Response<Long> getCommentCountByMember(@PathVariable Long memberId) {
-        Long count = commentService.countCommentsByMember(memberId);
+    public Response<Integer> getCommentCountByMember(@PathVariable Long memberId) {
+        int count = commentService.countCommentsByMember(memberId);
         return Response.success(count);
     }
